backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.123.9",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "pyjwt>=2.10.1",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "supabase>=2.25.0",
    "uvicorn[standard]>=0.38.0",
]
